api/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
pillow==11.1.0
rembg==2.0.57
slowapi==0.1.9
